Refactored Geo Queries
